+++ database/migrations/2021_09_14_142554_create_commentaires_table.php
@@ -16,9 +16,9 @@ class CreateCommentairesTable extends Migration
         Schema::create('commentaires', function (Blueprint $table) {
             $table->id();
             $table->tinyText('commentaire');
-            $table->unsignedBigInteger('user_id');
+            $table->integer('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->unsignedBigInteger('lieu_id');
+            $table->integer('lieu_id');
             $table->foreign('lieu_id')->references('id')->on('lieus')->onDelete('cascade');
             $table->integer('parent_id')->unsigned()->nullable();
             $table->timestamps();
